move mimetype hackery to content dropdown
