+++ packages/insomnia/src/models/__tests__/request.test.ts
@@ -2,7 +2,7 @@ import { beforeEach, describe, expect, it, jest } from '@jest/globals';
 
 import { globalBeforeEach } from '../../__jest__/before-each';
 import { CONTENT_TYPE_GRAPHQL } from '../../common/constants';
-import { newBodyGraphQL, updateMimeType } from '../../ui/components/panes/request-pane';
+import { newBodyGraphQL, updateMimeType } from '../../ui/components/dropdowns/content-type-dropdown';
 import * as models from '../index';
 
 describe('init()', () => {
